Migrated the validation logic out into extension methods
