+++ src/Retro.FastInject/Model/Manifest/ServiceManifest.cs
@@ -1,14 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
-using System.Text;
 using Microsoft.CodeAnalysis;
-using Microsoft.Extensions.DependencyInjection;
 using Retro.FastInject.Annotations;
 using Retro.FastInject.Comparers;
-using Retro.FastInject.Generation;
 using Retro.FastInject.Utils;
 
 namespace Retro.FastInject.Model.Manifest;
@@ -33,93 +29,26 @@ public class ServiceManifest {
   }
 
   /// <summary>
-  /// Validates the entire dependency graph for circular dependencies.
-  /// This should be called after all constructor dependencies have been resolved.
+  /// Adds a constructor resolution entry for a service type.
   /// </summary>
-  /// <exception cref="InvalidOperationException">Thrown when a circular dependency is detected.</exception>
-  public void ValidateDependencyGraph() {
-    var visited = new HashSet<ITypeSymbol>(TypeSymbolEqualityComparer.Instance);
-    var path = new Stack<ITypeSymbol>();
-    var onPath = new HashSet<ITypeSymbol>(TypeSymbolEqualityComparer.Instance);
-    
-    foreach (var serviceType in GetAllServices().Select(x => x.Type)) {
-      if (visited.Contains(serviceType)) continue;
-
-      if (DetectCycle(serviceType, visited, path, onPath, out var cycle)) {
-        throw new InvalidOperationException(
-            $"Detected circular dependency: {string.Join(" → ", cycle.Select(t => t.ToDisplayString()))}");
-      }
-    }
-  }
-  
-  private bool DetectCycle(ITypeSymbol type, HashSet<ITypeSymbol> visited, Stack<ITypeSymbol> path, 
-                          HashSet<ITypeSymbol> onPath, [NotNullWhen(true)] out List<ITypeSymbol>? cycle) {
-    cycle = null;
-    
-    
-    if (onPath.Contains(type)) {
-      return ExtractCycleFromPath(type, path, out cycle);
-    }
-    
-    if (!visited.Add(type)) {
-      return false; // Already visited and no cycle found
-    }
-
-    onPath.Add(type);
-    path.Push(type);
-    
-    // If we have a constructor resolution for this type, check its dependencies
-    if (_constructorResolutions.TryGetValue(type, out var resolution) && CheckServiceCycle(visited, path, onPath, ref cycle, resolution)) return true;
-
-    // Done with this node
-    path.Pop();
-    onPath.Remove(type);
-    return false;
-  }
-  private bool CheckServiceCycle(HashSet<ITypeSymbol> visited, 
-                                 Stack<ITypeSymbol> path, 
-                                 HashSet<ITypeSymbol> onPath, 
-                                 [NotNullWhen(true)] ref List<ITypeSymbol>? cycle, 
-                                 ConstructorResolution resolution) {
-    foreach (var serviceRegistration in resolution.Parameters
-                 .Select(param => new {
-                     param,
-                     isNullable = param.Parameter.Type.NullableAnnotation == NullableAnnotation.Annotated
-                 })
-                 .Where(t => !t.isNullable && t.param.DefaultValue == null)
-                 .Select(t => t.param.SelectedService)) {
-      // Check the selected service type if available
-      if (serviceRegistration is null) continue;
-
-      var serviceType = serviceRegistration.ResolvedType;
-      if (DetectCycle(serviceType, visited, path, onPath, out cycle)) {
-        return true;
-      }
-    }
-    return false;
-  }
-  private static bool ExtractCycleFromPath(ITypeSymbol type, Stack<ITypeSymbol> path, out List<ITypeSymbol> cycle) {
-    // We found a cycle
-    cycle = [];
-    var cycleStarted = false;
-      
-    // Extract the cycle from the path
-    foreach (var node in path.Reverse()) {
-      if (TypeSymbolEqualityComparer.Instance.Equals(node, type)) {
-        cycleStarted = true;
-      }
-        
-      if (cycleStarted) {
-        cycle.Add(node);
-      }
-    }
-      
-    cycle.Add(type); // Complete the cycle
-    return true;
-  }
-
+  /// <param name="resolution">The constructor resolution to be added, containing the type, constructor, and parameters.</param>
   public void AddConstructorResolution(ConstructorResolution resolution) {
     _constructorResolutions[resolution.Type] = resolution;
+  }
+
+  /// <summary>
+  /// Attempts to retrieve a constructor resolution for the specified type.
+  /// </summary>
+  /// <param name="type">The type symbol for which to retrieve the constructor resolution.</param>
+  /// <param name="resolution">
+  /// When this method returns, contains the constructor resolution associated with the specified type
+  /// if the resolution exists; otherwise, null.
+  /// </param>
+  /// <returns>
+  /// True if the constructor resolution for the specified type exists, otherwise false.
+  /// </returns>
+  public bool TryGetConstructorResolution(ITypeSymbol type, [NotNullWhen(true)] out ConstructorResolution? resolution) {
+    return _constructorResolutions.TryGetValue(type, out resolution);
   }
   
   /// <summary>
@@ -180,6 +109,12 @@ public class ServiceManifest {
     return GetAllServices().Where(reg => reg.Lifetime == lifetime);
   }
 
+  /// <summary>
+  /// Attempts to retrieve a list of service registrations for a given service type.
+  /// </summary>
+  /// <param name="serviceType">The type of the service to retrieve registrations for.</param>
+  /// <param name="services">Outputs the list of service registrations if found; otherwise, null.</param>
+  /// <returns>True if services were found for the given service type; otherwise, false.</returns>
   public bool TryGetServices(ITypeSymbol serviceType, [NotNullWhen(true)] out List<ServiceRegistration> services) {
     return _services.TryGetValue(serviceType, out services);
   }
